Create routes folder and move the routes.php
